Fix PostgreSQL type mismatches and bigint string returns
- Fix InitialSchema migration: shares→double precision, height/blockHeight→bigint,
  hashRate→double precision, telegramChatId→bigint, slotTime→bigint
- Fix entity definitions to match migration types
- Add missing hourlyStatsEnabled/hourlyWorkersEnabled migration for telegram
- Fix getRawMany bigint string→number conversions (slotTime Map key,
  blockHeight, bestDifficulty, external shares difficulty)
- Fix userAgent alias casing for PostgreSQL raw queries
- Fix ambiguous maxDifficulty column reference in ON CONFLICT upsert

+++ src/migrations/1717430400000-AddClientDifficultyStatistics.ts
@@ -16,7 +16,7 @@ export class AddClientDifficultyStatistics1717430400000 implements MigrationInte
                 "id" SERIAL NOT NULL,
                 "address" character varying(62) NOT NULL,
                 "clientName" character varying(64),
-                "slotTime" integer NOT NULL,
+                "slotTime" bigint NOT NULL,
                 "maxDifficulty" real NOT NULL DEFAULT '0',
                 CONSTRAINT "PK_client_difficulty_statistics_entity_id" PRIMARY KEY ("id")
             )
